Refactoring
Trying to clean up the Lists and make the intent more clear.
Removed the LinkedListArray, as that was just confusing and really
was only used for static storage for linked list nodes and not something
that was really needed. Test 01 still isn't passing, but test00 still
passes so we are still in a good place.