Correcting issues with the scheduler tests
Logic error with k_wait, a parent now returns -1 if there are no
children to wait for. The join has been updated to set its status
directly on the process that wants to join it. The entries in the PCB
are being reset by the process' parent prior to read.