Putting this back together.
Fixed issue with high CPU times, the dispatcher logic was wonky. It
should now be working correctly and not making unnecessary changes
thus reducing the CPU time on the Scheduler itself.
Tests 00-05 are all passing, there are some higher num tests here and
there that are also passing but the join logic still needs to be
implemented.