Removed the less generic original DoublyLinkedList
and replaced it with the my DoublySea DoublyLinkedList implementation
that uses generic void* pointers for all pointers and an offset to
determine where the data, pNext, and pPrev pointers are located in
the node structure. As long as they come in that order, the
DoublyLinkedNode can be replaced by any structure.