99.99% of all tests are passing.
The mailbox_free is working as expected. There are a few oddities with
the output for some of the tests, 4,9, and 25. BUT all tests function
correctly. With four, his output seems out of order, ours finishes
correctly for the priority of the processes. For test 25, our Ids are
different but they are both monotonically increasing.